exclusao da classe Banco, alteracoes na maisn usando a classe dados, finalizacao da criacao dos tipos de conta

+++ trabalho/src/ContaSalario.java
@@ -1,7 +1,7 @@
 public class ContaSalario extends Conta{
     private double limSaque, limTransacao;
 
-    public ContaSalario(int senha, int nroConta,Agencia ag,float saldo,Cliente cliente,double limTransacao,double limSaque)
+    public ContaSalario(int senha, int nroConta,Agencia ag,double saldo,Cliente cliente,double limTransacao,double limSaque)
     {
         super(senha,nroConta,ag,saldo,cliente);
